requirements-full.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
neo4j==5.16.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Geospatial
earthengine-api==0.1.384
geemap==0.30.4
geopandas==0.14.2
shapely==2.0.2
folium==0.15.1
rasterio==1.3.9
pyproj==3.6.1

# Data Processing
pandas==2.1.4
numpy==1.26.3
xarray==2024.1.1
netCDF4==1.6.5

# HTTP & APIs
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# Machine Learning
scikit-learn==1.4.0
joblib==1.3.2

# UI
streamlit==1.30.0
streamlit-folium==0.17.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
tenacity==8.2.3
